pill_detector/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.7
django-debug-toolbar==3.8.1
django-environ==0.10.0
django-extensions==3.2.1
djangorestframework==3.14.0
jupyter
pymysql==1.0.2
pytz==2022.7.1
ultralytics